pyproject.toml
[tool.pylint]
reportMissingImports = false
disable = "missing-function-docstring"

[tool.poetry]
name = "ssg"
version = "0.4.1"
description = "Github data for static site generators popularity"
authors = ["Evgeniy Pogrebnyak <e.pogrebnyak@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
pandas = "^1.4.2"
PyYAML = "^6.0"
pydantic = "^1.9.1"
requests-cache = "^0.9.4"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
myst-parser = "^0.18.0"
pytest = "^7.1.2"
Sphinx = "^5.0.2"
streamlit = "^1.10.0"
pyright = "^1.1.255"
pydata-sphinx-theme = "^0.9.0"
ghp-import = "^2.1.0"
pybadges = "^3.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
